separated viewer user controls and author controls
