+++ src/components/viewer/ViewerManager.tsx
@@ -147,6 +147,7 @@ export default class ViewerManager extends React.Component<Partial<IViewerManage
   }
 
   render() {
-    return this.props.visible && <Viewer/>;
+    // return this.props.visible && <Viewer/>;
+    return <Viewer/>;
   }
 }
